[refactor] BigInt -> Long 자료형 변경
JPA에서 Long으로 자료형을 사용하면 MySQL에 BigInt로 생성된다고 한다.

+++ src/main/java/com/myreviewservice/myreviewservice/dto/ReviewRequestDto.java
@@ -11,7 +11,7 @@ import java.math.BigInteger;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReviewRequestDto {
-    private BigInteger productId; //상품 ID
+    private Long productId; //상품 ID
     private String userid;
     private int score;
     private String content;
